+++  purenessscopeserver --username freeeyes1226@gmail.com/PurenessScopeServer/Proactor/ProConnectClient.cpp
@@ -188,7 +188,7 @@ bool CProConnectClient::SendData(ACE_Message_Block* pmblk)
 		}
 		else
 		{
-			nDebugSize = pmblk->length();
+			nDebugSize = (int)pmblk->length();
 		}
 
 		char* pData = pmblk->rd_ptr();
